do not re-allocate elemental matrices for diffusion in each gauss point

+++ src/pdes/mechanical/methods.h
@@ -50,7 +50,7 @@ extern int feenox_problem_bc_set_mechanical_traction(element_t *element, bc_data
 extern int feenox_problem_bc_set_mechanical_force(element_t *element, bc_data_t *bc_data, unsigned int v);
 
 // mechanical/bulk.c
-extern int feenox_problem_build_allocate_aux_mechanical(size_t n_nodes);
+extern int feenox_problem_build_allocate_aux_mechanical(unsigned int n_nodes);
 extern int feenox_problem_build_volumetric_gauss_point_mechanical(element_t *, unsigned int v);
 
 // material models
